Arguments for path added

+++ converter.py
@@ -1,15 +1,22 @@
 import csv
 import time
+import os
+import argparse
 import xml.etree.ElementTree as ET
 from tqdm import tqdm
-
-INPUT_FILE = "large_data.xml"
-OUTPUT_FILE = "data_converted.csv"
 
 CSV_HEADERS = ['id', 'name', 'category', 'price', 'stock']
 
 def convert_xml_to_csv(xml_file, csv_file):
-    print(f"Starting: {xml_file} -> {csv_file}")
+    if not os.path.exists(xml_file):
+        print(f"Error: File '{xml_file}' not found.")
+        return
+
+    file_size_mb = os.path.getsize(xml_file) / (1024 * 1024)
+    print("-" * 40)
+    print(f"Input file: {xml_file} ({file_size_mb:.2f} MB)")
+    print(f"Output file: {csv_file}")
+    print("-" * 40)
     start_time = time.time()
     
     with open(csv_file, 'w', newline='', encoding='utf-8') as f_out:
@@ -21,8 +28,6 @@ def convert_xml_to_csv(xml_file, csv_file):
         context = iter(context)
         _, root = next(context)
         
-        count = 0
-
         with tqdm(total=None, unit=" rec", desc="Processing") as pbar:
             for event, elem in context:
                 if elem.tag == "product":
@@ -34,7 +39,6 @@ def convert_xml_to_csv(xml_file, csv_file):
                         stock = elem.find('stock').text
                         
                         writer.writerow([p_id, name, category, price, stock])
-                        count += 1
                         pbar.update(1)
                     except AttributeError:
                         pass
@@ -44,7 +48,16 @@ def convert_xml_to_csv(xml_file, csv_file):
 
     end_time = time.time()
     duration = end_time - start_time
-    print(f"Success! Executed {count} items for {duration:.2f} seconds")
+    
+    print("-" * 40)
+    print(f"Success! Time of execution: {duration:.2f}/s")
 
 if __name__ == "__main__":
-    convert_xml_to_csv(INPUT_FILE, OUTPUT_FILE)
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument("input", help="Path to XML file")
+    parser.add_argument("output", help="Path to CSV file")
+
+    args = parser.parse_args()
+
+    convert_xml_to_csv(args.input, args.output)
